refactor UI components
