+++ web_src/src/pages/organization/settings/Integrations.tsx
@@ -349,9 +349,12 @@ export function Integrations({ organizationId }: IntegrationsProps) {
                                   if (!canUpdateIntegrations) return;
                                   const providerName = integration.metadata?.integrationName;
                                   if (providerName && integration.status?.setupState?.currentStep) {
-                                    navigate(`/${organizationId}/settings/integrations/${integration.metadata?.id}/setup`, {
-                                      state: { integrationId: integration.metadata?.id },
-                                    });
+                                    navigate(
+                                      `/${organizationId}/settings/integrations/${integration.metadata?.id}/setup`,
+                                      {
+                                        state: { integrationId: integration.metadata?.id },
+                                      },
+                                    );
                                     return;
                                   }
